requirements.txt
pillow
pillow-heif
boto3